Changed injector to handle inherited fields and non-fields
Inherited fields were not being injected as they did not belong to the root
class. To solve this we back track through the super classes and always
keep the root class.

Added methods to setText on labels and buttons.

Made the abstract inject method public which allows us to create labels and
inject data and not be limited to fields of an object